이구산업 - 20220127 재고실사 메뉴추가, 외주출고 중량 추가 및 수정

+++ app/src/main/java/kr/co/leeku/wms/network/MyDatabaseHelper.java
@@ -68,7 +68,7 @@ public class MyDatabaseHelper extends SQLiteOpenHelper {
         }
     }
 
-    public boolean insertData(int position, String barcode, String plt, float scan, String fg, String mac, int wg) {
+    public boolean insertData(int position, String barcode, String plt, float scan, String fg, String mac, float wg) {
         SQLiteDatabase db = this.getWritableDatabase();
         ContentValues contentValues = new ContentValues();
         contentValues.put(COLUMN_POSITION, position);
